Migrate 'user-interface-button-grouping' demo to NIM.
